+++ hw09_struct_validator/validators/minValidator.go
@@ -21,8 +21,8 @@ func MinValidator(requireMax string, v reflect.Value) error {
 			return ErrInvalidMin
 		}
 	case v.Kind() == reflect.Slice && v.Len() > 0 && slices.Contains(IntKinds, v.Index(0).Kind()):
-		for _, val := range v.Interface().([]int) {
-			if val < requireMaxVal {
+		for _, val := range v.Interface().([]int64) {
+			if val < int64(requireMaxVal) {
 				return ErrInvalidMin
 			}
 		}
